% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSKmod.R
\name{AIC}
\alias{AIC}
\title{AIC for Matrix-Variate Models}
\usage{
AIC(dist, Y, X, theta)
}
\arguments{
\item{dist}{Model distribution. Valid arguments are "mvgh", "mvvg", and "mvnig".}

\item{Y}{List of \eqn{n_i \times p} response matrices. Matrices must have same number of columns.}

\item{X}{List of \eqn{n_i \times q} design matrices. Matrices must have same number of columns.}

\item{theta}{List of parameter estimates}
}
\value{
AIC of the model.
}
\description{
Calculates AIC for MVGH, MVVG, and MVNIG models.
}
\examples{
AIC("mvvg", gaad_res, gaad_cov, mvvg_theta)
}
\author{
Samuel Soon

Dipankar Bandyopadhyay

Qingyang Liu
}
