% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSKmod.R
\name{BIC}
\alias{BIC}
\title{BIC for Matrix-Variate Models}
\usage{
BIC(dist, Y, X, theta)
}
\arguments{
\item{dist}{Model distribution. Valid arguments are "mvgh", "mvvg", and "mvnig".}

\item{Y}{List of \eqn{n_i \times p} response matrices. Matrices must have same number of columns.}

\item{X}{List of \eqn{n_i \times q} design matrices. Matrices must have same number of columns.}

\item{theta}{List of parameter estimates}
}
\value{
BIC of the model.
}
\description{
Calculates BIC for MVGH, MVVG, and MVNIG models.
}
\examples{
BIC("mvvg", gaad_res, gaad_cov, mvvg_theta)
}
\author{
Samuel Soon

Dipankar Bandyopadhyay

Qingyang Liu
}
