% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSKmod.R
\docType{data}
\name{gaad_cov}
\alias{gaad_cov}
\title{GAAD Data}
\format{
Each is a list of matrices, with rows denoting tooth sites and columns denoting CAL/PPD response.
}
\usage{
gaad_cov
}
\description{
These data sets describe periodontal measurements performed on members of the Gullah-Speaking African American community.
}
\details{
\code{gaad_res} and \code{gaad_cov} contain the response and covariate matrices of the GAAD data.
}
\examples{
gaad_cov
gaad_res
}
\keyword{datasets}
