% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSKmod.R
\name{MVSKmod}
\alias{MVSKmod}
\title{AECM Estimation for Matrix-Variate Skew Models}
\usage{
MVSKmod(Y, X, theta_g = NULL, dist, stopping = 0.001, max_iter = 50)
}
\arguments{
\item{Y}{List of \eqn{n_i \times p} response matrices. Matrices must have same number of columns.}

\item{X}{List of \eqn{n_i \times q} design matrices. Matrices must have same number of columns.}

\item{theta_g}{List of parameters to pass as initial values in the AECM algorithm.
If NULL, will be randomly generated. See Details for an in-depth explanation.}

\item{dist}{Modeling distribution. Select "mvgh" for matrix-variate generalized
hyperbolic, "mvvg" for variance-gamma, and "mvnig" for normal-inverse Gaussian.}

\item{stopping}{Stopping threshold for the L-infinity norm of differences in consecutive parameter space, evaluated at iteration \eqn{t+1} as
\eqn{|\hat\theta^{t+1} - \hat\theta^{t}|_\infty}. Default is 0.001}

\item{iter}{Maximum number of iterations, default is 50.}
}
\value{
MVSKmod returns a list with the following elements:

\itemize{
\item{\code{Iteration}: } {Number of iterations taken to convergence. \code{Inf} if convergence not reached.}
\item{\verb{Starting Value}: } {List of initial parameter values.}
\item{\verb{Final Value}: } {List of final parameter estimates.}
\item{\verb{Stopping Criteria}: } {Vector of \eqn{|\hat\theta^{t+1} - \hat\theta^{t}|_\infty} at each iteration.}
}
}
\description{
This function fits linear models for matrix data with varying row sizes,
according to the three matrix-variate skew distributions described in
(Gallaugher & McNicholas, 2019). Exchangeable observation row correlation
and skewness structures are imposed to accommodate the varying row counts
across matrices.
}
\details{
Fits the matrix-variate skew regression model

\deqn{Y_i = X_i \Theta + E_i,}

where each response \eqn{Y_i} is a \eqn{n_i \times p} matrix that indexes \eqn{n_i} observations and \eqn{p} response variables. \eqn{X_i} corresponds to a \eqn{n_i \times q} design matrix, and \eqn{\Theta} corresponds to a \eqn{q \times p} coefficient matrix. \eqn{E_i} corresponds to a \eqn{n_i \times p} error matrix, following a matrix-variate skew distribution.

Supported distributions include the matrix-variate generalized hyperbolic (MVGH), matrix-variate variance-gamma (MVVG), and matrix-variate normal-inverse Gaussian (MVNIG). Certain reparametrizations are implemented to avoid identifiability issues (Gallaugher & McNicholas, 2019). Additionally, the row correlation matrix \eqn{\Sigma_i} is reduced to a correlation matrix with exchangeable structure based on correlation parameter \eqn{\rho}, and skewness parameter \eqn{A_i} is represented as a function of skewness vector \eqn{a}: \eqn{A_i = \underline{1}_{n_i} a^T}.

The structure of \code{theta_g} and parameter estimates returned by the function must be in the form of a list with the following named elements:

\itemize{
\item{\code{Theta}: } {\eqn{q \times p} coefficient matrix}
\item{\code{a}: } {\eqn{p \times 1} skewness vector}
\item{\code{rho}: } {Compound symmetry parameter for row correlation matrix}
\item{\code{Psi}: } {\eqn{p \times p} column covariance matrix}
\item{\code{gamma}: } {Univariate mixing parameter, only included when running MVVG model}
\item{\code{tgamma}: } {Univariate mixing parameter, only included when running MVVG model}
}
}
\examples{
set.seed(1234)
# num response variables
p <- ncol(gaad_res[[1]])
# num covariates
q <- ncol(gaad_cov[[1]])
# generate initial value to input, then run AECM with MVVG distribution
initial_mvvg_theta <- list(Theta = matrix(rnorm(p*q), nrow = q, ncol = p),
                      A = rep(1,p),
                     rho = 0.3,
                     Psi = diag(p),
                     gamma = 4)
MVSKmod(gaad_res[1:50], gaad_cov[1:50], initial_mvvg_theta, "mvvg")

p <- ncol(gaad_res[[1]])
q <- ncol(gaad_cov[[1]])
initial_mvnig_theta <- list(Theta = matrix(rnorm(p*q), nrow = q, ncol = p),
                      A = rep(1,p),
                     rho = 0.3,
                     Psi = diag(p),
                     tgamma = 3)
MVSKmod(gaad_res[1:50], gaad_cov[1:50], initial_mvnig_theta, "mvnig")
}
\author{
Samuel Soon

Dipankar Bandyopadhyay

Qingyang Liu
}
