% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSKmod.R
\docType{data}
\name{mvvg_theta}
\alias{mvvg_theta}
\title{MVVG Parameter Format}
\format{
List of model parameters.
}
\usage{
mvvg_theta
}
\description{
This is an example of the format of input parameter list theta.
}
\examples{
mvvg_theta
}
\keyword{datasets}
